refactor: init server logic
